Struct political_economy_core::data_manager::DataManager

pub struct DataManager {

    pub sets: Vec<String>,
    pub wants: HashMap<usize, Want>,
    pub technology: HashMap<usize, Technology>,
    pub technology_families: HashMap<usize, TechnologyFamily>,
    pub products: HashMap<usize, Product>,
    pub skill_groups: HashMap<usize, SkillGroup>,
    pub skills: HashMap<usize, Skill>,
    pub processes: HashMap<usize, Process>,
    pub process_nodes: HashMap<usize, ProcessNode>,
    pub jobs: HashMap<usize, Job>,
    pub species: HashMap<usize, Species>,
    pub cultures: HashMap<usize, Culture>,
    pub territories: HashMap<usize, Species>,
    pub markets: HashMap<usize, Market>,
    pub pops: HashMap<usize, Pop>,
    pub firms: HashMap<usize, Firm>,
    /* private fields */

}

The DataManager is the main manager for our simulation It contains all of the data needed for the simulation in active memory, available for use as needed.

Items are stored and retrieved via their IDs. IDs for items should be fixed in place from loading onwards.

Required Items

Currently, there are a handful of items which are considered ‘Required’ by the system. These are items which should always be there and will effectively always have some use.

Required Wants - ID 0: Rest - TODO Items + Space + Free Space Required Products: - ID 0: Time (hr) (Produces 1 rest for owning it, made by pops at day start, refreshed every day) - ID 1: Shopping Time - TODO Items + Land (abstract) + Land (Wasteland) + Land (Marignal Land) + Land (Scrub Land) + Land (Quality Land) + Land (Fertile Land) + Land (Very Fertile Land) + Nothing (void item, may not be needed) Required Processes: - ID 0: Shopping (Time -> Shopping Time) Required Tech: - TODO Items + Brainstorming (origin tech)

Fields
sets: Vec<String>wants: HashMap<usize, Want>technology: HashMap<usize, Technology>technology_families: HashMap<usize, TechnologyFamily>products: HashMap<usize, Product>skill_groups: HashMap<usize, SkillGroup>skills: HashMap<usize, Skill>processes: HashMap<usize, Process>process_nodes: HashMap<usize, ProcessNode>jobs: HashMap<usize, Job>species: HashMap<usize, Species>cultures: HashMap<usize, Culture>territories: HashMap<usize, Species>markets: HashMap<usize, Market>pops: HashMap<usize, Pop>firms: HashMap<usize, Firm>
Implementations

impl DataManager

pub fn new() -> Self

pub fn load_wants(&mut self, _file_name: &String) -> Result<(), String>

Loads wants from a file into the data manager, Currently, this just loads pre-existing data.

pub fn load_technologies(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_technology_families(
&mut self,
_file_name: &String
) -> Result<(), String>

pub fn load_products(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_skill_groups(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_skills(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_processes(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_jobs(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_species(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_cultures(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_pop(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_territories(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_markets(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_firms(&mut self, _file_name: &String) -> Result<(), String>

pub fn load_all(&mut self, _file_name: &String) -> Result<(), String>

Placeholder loader for everything Should load by sets later on, rather than all at once.

impl DataManager

pub fn new_want_id(&mut self) -> usize

pub fn new_tech_id(&mut self) -> usize

pub fn new_tech_fam_id(&mut self) -> usize

pub fn new_product_id(&mut self) -> usize

pub fn new_skill_group_id(&mut self) -> usize

pub fn new_skill_id(&mut self) -> usize

pub fn new_process_id(&mut self) -> usize

pub fn new_job_id(&mut self) -> usize

pub fn new_species_id(&mut self) -> usize

pub fn new_culture_id(&mut self) -> usize

pub fn new_pop_id(&mut self) -> usize

pub fn new_territory_id(&mut self) -> usize

pub fn new_market_id(&mut self) -> usize

pub fn new_firm_i(&mut self) -> usize

impl DataManager

This impl block contains no items.

Sanity check functions, ensures no duplicate names and that the items loaded equals the items in memory.