Struct political_economy_core::objects::desire::Desire

pub struct Desire {
    pub item: DesireItem,
    pub start: u64,
    pub end: Option<u64>,
    pub amount: f64,
    pub satisfaction: f64,
    pub reserved: f64,
    pub step: u64,
    pub tags: Vec<DesireTag>,
}

Desires

Desires are things that are desired and used in a the Desires class. It contains an item, either a want or product, and at minimum a tier at which it applies.

This will likely be broken up and simplified in the future, distinguishing between recorded desires, such as the desires of species, Cultures, etc, and working desires, which are used for satisfaction and trade guarantees.

Fields
item: DesireItemstart: u64end: Option<u64>amount: f64satisfaction: f64reserved: f64step: u64tags: Vec<DesireTag>
Implementations

impl Desire

pub fn is_match(&self, other: &Desire) -> bool

Checks if a desire is a match for the current desire.

It matches on the

item (product/Want and Id)

Start

End

Step

No other factors are used. This is used to make adding desires together easier.

Currently tags are not used. Be aware when using this.

pub fn add_satisfaction(&mut self, add: f64) -> f64

Adds the given value to our satisfaction, capping at our maximum and returning the excess.

Not meant for negative values, but they are accepted.

pub fn satisfaction_up_to_tier(&self) -> Option<u64>

Calculates what tier this desire is satisfied to, stopping at the last tier that has any satisfaction in it. If it fully fills it’s highest tier it returns that tier.

IE. Start = 1, Step size 1, amount / level 1.0

Satisfaction = 100. Highest level satisfied = 100 (not 101).

If the satisfaction is 0, it returns None.

pub fn get_next_tier_up(&self, tier: u64) -> Option<u64>

Get Next Tier up takes a given tier and gets the next valid tier up.

Returns err if no tier is next. Otherwise, returns the next valid tier.

pub fn change_end(&mut self, end: Option<u64>, step: u64) -> Result<(), String>

Changes the end and step for the desire.

pub fn steps(&self) -> u64

Gets how many steps the desire has in total.

If the desire is infinite, it return 0.

If the desire has only 1 tier, then it’s 1.

pub fn is_stretched(&self) -> bool

If the Desire covers more than 1 tier.

pub fn is_infinite(&self) -> bool

If the desire is infinite, with no end tier.

pub fn satisfaction_at_tier(&self, tier: u64) -> Result<f64, String>

How much satisfaction this desire has at a specific tier.

Returns the amount in units requested satisfied at this level. It caps at 0 and self.amount.

pub fn steps_to_tier(&self, tier: u64) -> Result<u64, String>

How many steps it takes for this desire to reach the given tier.

pub fn steps_on_tier(&self, tier: u64) -> bool

Checks whether a given value steps on a tier of this desire.

Examples

use political_economy_core::objects::desire::Desire;
use political_economy_core::objects::desire::DesireItem;

let desire = Desire{item: DesireItem::Product(0),start: 2, end: Some(10),
    amount: 5.0, satisfaction: 15.0, reserved: 0.0, step: 2, tags: vec![]};
assert_eq!(desire.steps_on_tier(4), true);

pub fn is_fully_satisfied(&self) -> bool

Checks if the desire is fully satisfied or not.

pub fn total_desire(&self) -> Option<f64>

Retrieves the total amount desired across all tiers for the desire.

If the desire is infinite, then it returns -1.

pub fn total_satisfaction(&self) -> f64

Calculates the total satisfaction of the desire, returning it measured in the number of tiers satisfied.

Does not include skipped tiers

pub fn total_desire_at_tier(&self, tier: u64) -> Result<f64, String>

Gets the total amonut desired up to the tier given.

If the tier is below the starting tier it returns 0.

If the tier is between its start and end tier, it returns the amount * the tiers belowe it’s current tier.

pub fn unsatisfied_to_tier(&self) -> Option<u64>

Walking up the desire, it gets the first tier which has any unsatisfied quantity.

If totally sastisfied at a tier, it will return the next available tier, or None.

If it is totally satisfied, it returns None.