Struct political_economy_core::objects::desires::Desires

pub struct Desires {
    pub desires: Vec<Desire>,
    pub property: HashMap<usize, f64>,
    pub want_store: HashMap<usize, f64>,
    pub shopping_data: HashMap<usize, DesireInfo>,
}

Desires are the collection of an actor’s Desires. Includes their property excess / unused wants, and AI data for acting on buying and selling.

Fields
desires: Vec<Desire>

All of the desires we are storing and looking over.

property: HashMap<usize, f64>

The property currently owned bey the actor.

want_store: HashMap<usize, f64>

The wants stored and not used up yet.

shopping_data: HashMap<usize, DesireInfo>

The data of our succes in shopping.

Implementations

impl Desires

pub fn new(desires: Vec<Desire>) -> Self

Creates a new desire collection based on a list of desires.

pub fn sift_products(&mut self)

Goes over the property contained within desires and sifts them into the various desires that need them.

pub fn sift_product(&mut self, product: &usize) -> Result<(), DesireError>

Sifts a singular product into the various desires that seek it.

Plan

This will need to be expanded to allow for both specific product satisfaction as well as general product satisfaction.

pub fn add_property(&mut self, product: usize, amount: &f64)

Adds a number of units to the property.

pub fn add_want(&mut self, want: usize, amount: &f64)

Adds a number of wants.

pub fn add_desire(&mut self, desire: &Desire)

Adds a desire to our collection.

If the desire is a match for an existing desire, it will combine it with the existing match, adding the amount desired per level.

This can also be used to subtract a desire from the collection.

If a desire is reduced to 0 or less, then it will be removed.

pub fn barter_value_difference(
&self,
items_in: Vec<(usize, f64)>,
items_out: Vec<(usize, f64)>
) -> ValueInOut

Gets the difference in value between the items in and the items out given the current state of the desires.

If items are on an equal tier then they are equal on a unit-per-unit basis. For each tier of separation the value diverges geometrically. For every tier of separation the value of an item declines by 0.9^(n) per unit, where n is the difference between the lower and higher tier.

If item_out is not currently desired (either having no desires which want it or a desire which matches, but it is totally satisfied) then it’s value is equal to 0.

It returns the Values in and out, letting the caller use that information at their leasure.

pub fn out_barter_value(&self, product: &usize, amount: f64) -> Option<(u64, f64)>

Retrieves the internal barter value for a specific product as though it were going out and becoming unable to satisfy desires.

It returns the tier at which it starts, and the internal satisfaction value it will lose.

If there is no desire for that product exists, or there is no satisfaction to pull from, it returns None.

Example

A product is desired at at tiers 0 2 4 and 6. One unit each. and it has 2 units of satisfaction.

We want to remove 1.5 units of the item.

It finds that desire and sets the tier to 2, it then goes through, pretending to subtract and count up the satisfaction lost. In this case it has 1 from 2 and 0.5 * 0.9^-2 for a total loss of roughly 1.61728.

The resulting output would be Some(2, 1.61728).

pub fn get_highest_satisfied_tier_for_item(
&self,
item: DesireItem
) -> Option<u64>

Finds the highest tier for a particular item which has satisfaction

Returns None if no satisfaction in any product found.

pub fn get_highest_satisfied_tier(&self) -> Option<u64>

Finds the highest tier which has any satisfaction.

Returns None if no satisfaction in any product found.

pub fn walk_down_tiers(&self, prev: &DesireCoord) -> Option<DesireCoord>

Take in a location and walk from that location back down our tiers. This should result in a full reversal of self.walk_up_tiers().

This cannot accept a None value, as desires can be infinite, creating no guaranteed that there is a ‘last’ desire.

The Desire Coord does not need to have an idx < self.desires.len(), if it has an idx > self.desires.len() it will drag it down to len, allowing you to start at a tier specifically, without touching a higher tier.

pub fn walk_down_tiers_for_item(
&self,
prev: &DesireCoord,
item: &DesireItem
) -> Option<DesireCoord>

Like walk_down_tiers(), this walks down the desires we have. But, it only returns the indexes and tiers for a specific desire, instead of all of them.

This cannot accept a None value, as desires can be infinite, creating no guaranteed that there is a ‘last’ desire.

The Desire Coord does not need to have an idx < self.desires.len(), if it has an idx > self.desires.len() it will drag it down to len, allowing you to start at a tier specifically, without touching a higher tier.

pub fn get_lowest_unsatisfied_tier_of_item(
&self,
item: DesireItem
) -> Option<u64>

Take an item and finds the lowest tier available which can still accept the item.

Used primarily to nicely find where to put an item when sifting.

pub fn get_lowest_unsatisfied_tier(&self) -> Option<u64>

Finds the lowest unsatisfied tier of all of our desires.

If all desires are satisfied, it returns None.

pub fn in_barter_value(&self, product: &usize, amount: f64) -> Option<(u64, f64)>

Retrieves the internal barter value for a specific product as though it were coming in and satisfying desires.

It returns the tier at which it starts, and the internal satisfaction value it satisfies.

If it is not desired at all, it returns None.

Example

A product is desired at at tiers 5 7 9 and 11. One unit each.

It already has 1 unit in it (tier 5), and we want to insert 1.5 units more.

It finds that desire and sets the tier to 7, it then walks up the desire and calculates the weight (1 + 0.5*0.9 ≅ 1.45).

The resulting output would be Some(7, 1.45).

pub fn tier_equivalence(start: u64, end: u64) -> f64

Tier Equivalence between two tiers.

With the start tier equal to 1, end tiers above it decline by 0.9 per level difference. Going down they increase by 1/0.9

This defines how many units at the end tier is considered equivalent to lose in return for 1 unit of the start tier.

IE

start 10, end 11 = 0.9^-1 1 start = 0.9 end

start 10, end 12 = 0.9^-2 1 start = 0.81 end

start 10, end 8 = 0.9^2 1 start = 1.23. end

pub fn walk_up_tiers(&self, prev: Option<DesireCoord>) -> Option<DesireCoord>

Walk up the tiers of our desires.

It takes as a parameter our previous desire location. If given none, it selects the first and return that (lowest tier, first in vector order).

If there is no next step, it returns None.

pub fn walk_up_tiers_for_item(
&self,
prev: &Option<DesireCoord>,
item: &DesireItem
) -> Option<DesireCoord>

Walk up the tiers for a specific desire item.

It takes in the tier we are starting at, and the index of the previous we looked at. The previous index need not be valid to be used.

If there is no next step, it returns None.

pub fn clear_desires(&mut self)

Clears self.desires

pub fn len(&self) -> usize

How many desires are contained.