Struct political_economy_core::objects::pop_breakdown_table::PopBreakdownTable

pub struct PopBreakdownTable {
    pub table: Vec<PBRow>,
    pub total: usize,
}

Pop Breakdown Table

Stores the information and ‘rows’ of a population’s data table. Each

Fields
table: Vec<PBRow>

The table of all our data, broken up by the row’s categories. Species(Cohort, Subtype), Culture(Class, Generation), Ideology(Wave, Faction)

total: usize

The sum of all the people in the table. Keep updated with any change.

Implementations

impl PopBreakdownTable

pub fn insert_pops(&mut self, row: PBRow)

Inserts a pop into the table. If a row in the table matches it’s demographics, then add to that pop instead of a new row.

pub fn remove_pops(&mut self, row: &PBRow)

Removes pops from the table. If this empties the row, then it removes that row.

pub fn species_makeup(&self) -> HashMap<usize, usize>

Gets the number of pops in each species.

pub fn culture_makeup(&self) -> HashMap<Option<usize>, usize>

Gets the number of pops in each species.

pub fn ideology_makeup(&self) -> HashMap<Option<usize>, usize>

Gets the number of pops in each species.

pub fn species_division(&self) -> HashMap<usize, f64>

Counts up and returns the percentile division into each species.

pub fn culture_division(&self) -> HashMap<Option<usize>, f64>

Counts up and returns the percentile division into each culture. Includes pops who have no Ideology.

pub fn ideology_division(&self) -> HashMap<Option<usize>, f64>

Conuts up and returns the percentile division into each ideology. Includes pops who have no Ideology.