Enum political_economy_core::objects::product::ProductTag

pub enum ProductTag {

    SelfLuxury {
        efficiency: f64,
    },
    WantLuxury {
        want: usize,
        efficiency: f64,
    },
    ProductLuxury {
        product: usize,
        efficiency: f64,
    },
    SelfBargain {
        efficiency: f64,
    },
    WantBargain {
        want: usize,
        efficiency: f64,
    },
    ProductBargain {
        product: usize,
        efficiency: f64,
    },
    Public {
        capacity: usize,
        decay: f64,
    },
    Share {
        firm: usize,
    },
    Animate,
    ConsumerGood,
    MilitaryGood,
    Fixed,
    Currency,
    Service,
    Remote,
    Invariant,
    Abstract,
    Liquid,
    Gas,
    Atomic {
        protons: u32,
        neutrons: u32,
    },
    Storage {
        storage_type: StorageType,
        mass: f64,
        volume: f64,
    },
    Vulnerable {
        to: ProductVulnerability,
    },
    Magic,

}

Variants

SelfLuxury

Fields

efficiency: f64

The item has improved efficiency at satisfying itself as it’s price increases.

WantLuxury

Fields

want: usize

efficiency: f64

The item has improved efficiency at satisfying a want as it’s price increases.

ProductLuxury

Fields

product: usize

efficiency: f64

The item has improved efficiency at satisfying a different product as it’s price increases.

SelfBargain

Fields

efficiency: f64

The item has improved efficiency at satisfying itself as it’s price decreases.

WantBargain

Fields

want: usize

efficiency: f64

The item has improved efficiency at satisfying a want as it’s price decreases.

ProductBargain

Fields

product: usize

efficiency: f64

The item has improved efficiency at satisfying a different product as it’s price decreases.

Public

Fields

capacity: usize

decay: f64

The item is a public good, the value it contains is how many people it can be shared between without losing efficiency. Decay is how quickly it falls per person after it maxes out.

Share

Fields

firm: usize

The product is a share of ownership in a company. The company manages the details of the share’s effects on the firms

Animate

The item is actually a living creature (placed her for later purposes.)

ConsumerGood

The item is a consumer good, making variants is more open ended, but the product cannot be used as an input in a process. Only Failure processes are allowed.

MilitaryGood

The item is a military good, this gives it more flexibility in variation, but removes it as an input or capital processes. The only processes it allows are failure processes.

Fixed

The item is immobile, it can only be traded within a market or territory, and cannot be transferred out of a market or territory.

Currency

The item is a currency, marking it as the preferred item to use to buy or sell items.

Service

The item is a service, it fails each day, turns into nothing, and cannot be transferred outside of the market.

Remote

The item is a service, but can be sold outside of the market via communication methods.

Invariant

The item cannot be made distinguishable from any other and cannot have variants.

Abstract

The item is an abstract ideal, it can have vairants, but not a ‘default’.

Liquid

The item is a liquid, requiring special storage to keep. It also acts as a liquid once it enters the environment.

Gas

The item is a gas, requiring special storage to keep. It also acts as a gas once it enters the environment.

Atomic

Fields

protons: u32

neutrons: u32

The item is a pure atomic material and is thus available for matter smasher stuff.

Storage

Fields

storage_type: StorageType

mass: f64

volume: f64

The item can store other items in it. It defines what kind of storage it offers, and the amount it can store interms of mass and valume. Storage negates or reduces degredation effects from the environment and vulnerability.

Vulnerable

Fields

to: ProductVulnerability

The product has particular vulnerabilities,

Magic

The item is magical, and thus can break our normal rules and requirements for an item.