Struct political_economy_core::objects::product::Product

pub struct Product {

    pub name: String,
    pub variant_name: String,
    pub description: String,
    pub unit_name: String,
    pub quality: i32,
    pub mass: f64,
    pub bulk: f64,
    pub mean_time_to_failure: Option<u32>,
    pub fractional: bool,
    pub tags: Vec<ProductTag>,
    pub wants: HashMap<u64, f64>,
    pub processes: HashSet<u64>,
    pub failure_process: Option<u64>,
    pub use_processes: HashSet<u64>,
    pub consumption_processes: HashSet<u64>,
    pub maintenance_processes: HashSet<u64>,
    pub tech_required: Option<u64>,
    /* private fields */

}

Fields
name: Stringvariant_name: Stringdescription: Stringunit_name: Stringquality: i32mass: f64bulk: f64mean_time_to_failure: Option<u32>fractional: booltags: Vec<ProductTag>wants: HashMap<u64, f64>processes: HashSet<u64>failure_process: Option<u64>use_processes: HashSet<u64>consumption_processes: HashSet<u64>maintenance_processes: HashSet<u64>tech_required: Option<u64>
Implementations

impl Product

pub fn new(
id: u64,
name: String,
variant_name: String,
description: String,
unit_name: String,
quality: i32,
mass: f64,
bulk: f64,
mean_time_to_failure: Option<u32>,
fractional: bool,
tags: Vec<ProductTag>,
tech_required: Option<u64>
) -> Option<Self>

pub fn is_equal_to(&self, other: &Product) -> bool

Checks if a product is overly similar to another. Not Thoroughly tested.

pub fn get_name(&self) -> String

pub fn set_want(&mut self, want: &Want, eff: f64) -> Result<(), &str>

Sets the value of a want that this product satisfies via ownership

pub fn connect_want(&mut self, want: &mut Want, eff: f64) -> Result<(), &str>

As set_want(self, want, eff), but also ensures that want is connected back.

pub fn add_tag(&mut self, tag: ProductTag)

pub fn get_tags(&self, tag: &ProductTag) -> Vec<&ProductTag>

!TODO Test this for correctness and if there is a better way to do it.

pub fn id(&self) -> u64

pub fn add_process<'a>(&mut self, process: &Process) -> Result<(), &'a str>

Adds a process to the product. Also adds it to all appropriate subcategories. Returns a Err if duplicate failure product was found.